xprt_rdma: Fix up xprt_create_id()
This patch brings xprt_create() to a point where it might
be able to get an endpoint of an RDMA conenction.

The "struct xprt_rdma_ep *" passed in will be filled in with the
"id", which in turn has a "ib_qp *" in it, when successful.

It returns 0 upon success and non-zero for errors.